+++ packages/cache/src/CacheStore.spec.tsx
@@ -1,4 +1,4 @@
-import { ERROR_MESSAGE, FALLBACK, TEXT, sleep } from '@suspensive/test-utils'
+import { ERROR_MESSAGE, FALLBACK, TEXT, sleep } from '@suspensive/utils'
 import { render, screen, waitFor } from '@testing-library/react'
 import ms from 'ms'
 import { Suspense } from 'react'