+++ packages/react/src/ErrorBoundary.test-d.tsx
@@ -1,7 +1,6 @@
-import { CustomError, CustomNotError } from '@suspensive/test-utils'
+import { type ConstructorType, CustomError, CustomNotError } from '@suspensive/utils'
 import type { ComponentProps, ReactNode } from 'react'
 import { ErrorBoundary } from './ErrorBoundary'
-import type { ConstructorType } from './utility-types'
 
 describe('<ErrorBoundary/>', () => {
   it('should pass only boolean or ErrorConstructor or ShouldCatchCallback or ShouldCatch[]', () => {