+++ packages/react/src/utils/index.ts
@@ -1,3 +1,2 @@
 export { hasResetKeysChanged } from './hasResetKeysChanged'
-export { increase } from './increase'
-export { noop } from './noop'
+export { defineSuspense } from './defineSuspense'