+++ packages/react-query-4/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @suspensive/react-query-4
 
+## 2.9.0
+
+### Patch Changes
+
+- Updated dependencies [[`f9fb719`](https://github.com/toss/suspensive/commit/f9fb719c65a8de198946ac75933a8d9130ff1ce9)]:
+  - @suspensive/react@2.9.0
+  - @suspensive/utils@2.9.0
+
+## 2.8.1
+
+### Patch Changes
+
+- Updated dependencies [[`2326b13`](https://github.com/toss/suspensive/commit/2326b1341f167454a889953fb0bbf58449e1ca98)]:
+  - @suspensive/utils@2.8.1
+  - @suspensive/react@2.8.1
+
 ## 2.8.0
 
 ### Patch Changes